images/trustx-cml-firmware.bb: fixed location of files
The firmware image contained an folder named firmware with all
the firmware files. However this is mounted on '/lib/firmware'.
Thus we had to move the files to the root directory of the firmware image.
Further, the regulatory.db file was not built into the image before.
This is also fixed now.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>